clean up index(s) and Portrait fill in fix
