+++ .gitignore
@@ -3,3 +3,4 @@
 public/dist/static
 node_modules
 npm-debug.log
+languages.txt
