+++ docs/index.html
@@ -118,7 +118,7 @@
 	</div>
 
 	<footer>
-		<a href="https://github.com/kylebutts/phone_your_rep">
+		<a href="https://github.com/phoneyourrep/phone_your_rep">
 			<i class="fa fa-github font-white icon" aria-hidden="true"></i>
 		</a>
 	</footer>
